[-] Ongeki: Remove unused models
